backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
torch>=1.7.0
torchvision>=0.8.1
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pyyaml==6.0.1
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.3
scipy==1.11.4
